pyproject.toml
[project]
name = "ftm-similarity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "sentence-transformers>=2.2.2",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
]
